remove http service file
