+++ http/svc_hammer.go
@@ -9,6 +9,19 @@ import (
 	"github.com/kudarap/dotagiftx"
 )
 
+// hammerService provides access to hammer service methods used by http handlers.
+type hammerService interface {
+	// Ban updates user status to ban and cancels all listings.
+	Ban(context.Context, dotagiftx.HammerParams) (*dotagiftx.User, error)
+
+	// Suspend updates user status to suspend and cancels all listings.
+	Suspend(context.Context, dotagiftx.HammerParams) (*dotagiftx.User, error)
+
+	// Lift update user status to "marked" and remove its ban or suspend a flag
+	// and will restore items if requested.
+	Lift(ctx context.Context, steamID string, restoreListings bool) error
+}
+
 func handleHammerBan(svc hammerService, cache cacheManager, logger *slog.Logger) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		var p dotagiftx.HammerParams
