+++ http/svc_user.go
@@ -1,6 +1,7 @@
 package http
 
 import (
+	"context"
 	"fmt"
 	"log/slog"
 	"net/http"
@@ -12,6 +13,39 @@ import (
 )
 
 const userCacheExpr = time.Minute * 5
+
+// userService provides access to user service methods used by http handlers.
+type userService interface {
+	// FlaggedUsers returns a list of flagged/reported users.
+	FlaggedUsers(ctx context.Context, opts dotagiftx.FindOpts) ([]dotagiftx.User, error)
+
+	// User returns user details by id.
+	User(ctx context.Context, id string) (*dotagiftx.User, error)
+
+	// UserFromContext returns user details from context.
+	UserFromContext(context.Context) (*dotagiftx.User, error)
+
+	// CreateSubscription creates a subscription for the current user.
+	CreateSubscription(ctx context.Context, planID string) (subscriptionID string, err error)
+
+	// ProcessSubscription validates and processes subscription features.
+	ProcessSubscription(ctx context.Context, subscriptionID string) (*dotagiftx.User, error)
+
+	// UpdateSubscriptionFromWebhook handles user subscription updates from http request.
+	UpdateSubscriptionFromWebhook(ctx context.Context, r *http.Request) (*dotagiftx.User, error)
+
+	// ProcessManualSubscription processes manual subscription.
+	ProcessManualSubscription(ctx context.Context, form dotagiftx.ManualSubscriptionParam) (*dotagiftx.User, error)
+}
+
+// steamClient provides access to steam API methods used by http handlers.
+type steamClient interface {
+	// Player returns player summary base on steamID.
+	Player(steamID string) (*dotagiftx.SteamPlayer, error)
+
+	// ResolveVanityURL returns steam id from profile url.
+	ResolveVanityURL(url string) (steamID string, err error)
+}
 
 func handleProfile(svc userService, cache cacheManager, logger *slog.Logger) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
